+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased]
+
+### Added
+- private class methods on `Performance` & `Cinema`
+- improved documentation
+
+
 ## 3.0.1 - 2016-02-03
 
 ### Removed